Add files via upload
This is the code that uses the pre-processed data from the cookstove sensors to analyse cookstover performance. Infrared (EXACT) sensors are installed on the three stone fire stoves. Thermocouple (EXACTv2) sensors are installed on the mudstoves, charcoal and impoved biomass stoves. Fuel consumtion (FUEL) sensors are installed on the fuel baskets. Air quality (HAPEx) sensors are installed in the kitchens.